MLOps/code/05/requirements.txt
sklearn
scipy
numpy
pandas
matplotlib
azureml-sdk